whisper-server/requirements.txt
# Whisper Server Requirements
openai-whisper==20231117
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
kombu==5.3.2
aiofiles==23.2.1
scipy==1.11.4